Project Movie
A Simple Project which I developed While learning Python. It consists of
a frontend desingned using HTML and Javascript. The HTML page consists
of posters for different movies, when you click on that poster , The
trailer of that particular movie is played.The entire backend is
developed in Python.

In Order to run the Project

1. Clone the folder Project _Movie to your system.

2. From Your System open the file fresh_tomatoes.html from your browser.
OR
you can also run the file "entertainment_media.py" from your command
line it will open the HTML page for you

3. Click on any Movie poster to play their trailer.