Code Breaker
It's  a simple command line game so put together , The game goes like
this:

1. The computer will think of 3 digit number that has no repeating
digits.
2. You will then guess a 3 digit number
3. The computer will then give back clues, the possible clues are:

Close: You've guessed a correct number but in the wrong position

Match: You've guessed a correct number in the correct position

Nope: You haven't guess any of the numbers correctly

4. Based on these clues you will guess again until you break the code
with a
perfect match, the game will report "CODE CRACKED"!