Lift Management System
A 30- storeyed building has got about 5 wings where there would be a
lift in each of the wing. You have to make available a lift to the
person who presses a button to
get a lift. Follow the steps given below for designing the program.

1. The floors of the building should be numbered as 0 - 29.
2. The lifts should be numbered as 0 - 4.
3. Display a menu that would have following options
a. Do you wish to use a lift?
b. Show lift status
c. Exit
If user selects, option 1 then get following information from the user.

-Get the floor number where the person is standing
-Whether user wishes to go up/down.
-Get the floor number where the user wishes to go.

The validation checks should be there for the data that user enters.
The validation checks are as given below.
1. The floor number where the user is standing and the floor number
where user wishes to go should be in a range of 0 to 29.
2. If the user is standing on ground floor, i.e. 0th floor, then
selecting the direction for the lift to go down would be invalid.
3. If the user is standing on topmost floor, i.e. 29th floor, then
selecting the direction for the lift to go up would be invalid.
4. The direction for the lift whether up or down should be entered as
'u' or 'd' only.